configs for mixing, deduping, and tokenizing steps for reddit ablation baseline
The filtering of long, short, and contaminated files was done to
produces the "a0" mix. The settings used for this step are in mixer_config_baseline.json.

This a0 mix formed the base for all reddit-related ablation runs. The a1 mix is a 30%
random sample of a0 and is used in the reddit-v1-ablation-base_3645017
run.