Add ability to restart on new epoch
You can set the epoch via the option `--epoch=[INTEGER]`. This
automatically handles changing the data order each epoch by setting the
data seed to `seed + epoch`. So `--epoch` is the only flag you need to
set when restarting on a new epoch. Everything else in the config can
stay the same.

Note that we count epochs starting from 0. So to start the 2nd epoch you would add
the flag `--epoch=1`.